v0.0.5: Time to stop languishing. Here's a redo with tests
Removed `subcp` because I never used it. That makes this app focused
solely on move/rename operations, which simplifies it greatly (though
the old code could still probably have been simplified).

Removed log level options because I never used them.

Removed `--trial` because my usage pattern was almost always to
preview the operation, then run it. `--trial` would just force a
rollback, with its intent to be like a dry run that is actually
informed by the rules of the OS and FS.